Move relativeTimestampToString to utils so it can be used elsewhere
